+++ app/soc/PE_Types.h
@@ -31,7 +31,7 @@ typedef unsigned long int   VUINT32;
 typedef signed char int8_t;
 #endif
 #ifndef int16_t
-typedef signed int   int16_t;
+typedef signed short   int16_t;
 #endif
 #ifndef int32_t
 typedef signed long int    int32_t;
@@ -41,7 +41,7 @@ typedef signed long int    int32_t;
 typedef unsigned char       uint8_t;
 #endif
 #ifndef uint16_t
-typedef unsigned int  uint16_t;
+typedef unsigned short  uint16_t;
 #endif
 #ifndef uint32_t
 typedef unsigned long int   uint32_t;
